pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lakehouse-iot-telemetry-pipeline"
version = "1.0.0"
description = "Multi-tenant IoT telemetry Lakehouse on Databricks + Delta Lake"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Sushma K L", email = "sushmakl95@gmail.com" },
]
keywords = ["databricks", "delta-lake", "pyspark", "lakehouse", "iot", "data-engineering"]

dependencies = [
    "pyspark==3.5.1",
    "delta-spark==3.2.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "structlog>=24.1.0",
    "click>=8.1.7",
    "faker>=24.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-spark>=0.6.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "bandit>=1.7.8",
    "detect-secrets>=1.5.0",
    "pre-commit>=3.7.0",
    "chispa>=0.10.1",
]
databricks = [
    "databricks-sdk>=0.24.0",
    "databricks-cli>=0.18.0",
]

[project.scripts]
iot-pipeline = "src.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 110
target-version = "py311"
extend-exclude = ["notebooks/*", "data/*"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "S",    # bandit security
    "RUF",  # ruff-specific
]
ignore = ["S101", "E501"]  # allow assert in tests, long lines in SQL

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing"
testpaths = ["tests"]
markers = [
    "unit: fast unit tests",
    "integration: integration tests requiring spark + delta",
    "slow: tests that take >5s",
]

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["*/tests/*", "*/notebooks/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
